Fix prompt on windows (#2986)
* Fix typo

* Refactor prompt highlighting into RLHelper

By moving the prompt coloring to be done in
Highlighter::highlight_prompt, we sidestep the issue on Windows where
the prompt width is calculated post-coloring AND without ignoring escape
codes.

By including it in the implementation of Highlighter, Editor::readline
now operates on a plain-text prompt, so width calculation is correct.

This commit also re-arranges the trait impl order to match the
definition.